testproj_with_docker/requirements.txt
djangorestframework
redisearch[hiredis]
django_redis